requirements.txt
Flask==2.3.0
Werkzeug==2.3.0
openai==0.27.8
python-dotenv==1.0.0
chromadb==0.4.15
sentence-transformers==2.2.2
PyPDF2==3.0.1
python-docx==0.8.11
openpyxl==3.1.2